# Micro-optimization for `IsPrimitiveType`